refactor(cli): drive query commands through a declarative command table
