+++ cli/src/cli.ts
@@ -9,58 +9,30 @@ import {
   analyseCase,
   exportCase,
   ingestSource,
-  queryAutofill,
-  queryBookmarks,
-  queryCache,
-  queryCandidates,
-  queryCookies,
-  queryCredentials,
-  queryDownloads,
-  queryFavicons,
-  queryHistory,
-  queryMetadata,
-  queryTopicCandidates,
-  queryTopSites,
   renderReport,
-  type AutofillDirection,
-  type AutofillSort,
-  type CandidateDirection,
-  type CandidateSort,
-  type BookmarkDirection,
-  type BookmarkSort,
-  type BookmarkSource,
-  type CacheDirection,
-  type CacheRecordType,
-  type CacheSort,
   type CommitState,
-  type CookieDirection,
-  type CookieSort,
-  type CredentialDirection,
-  type CredentialSort,
-  type DownloadDirection,
-  type DownloadSort,
-  type FaviconDirection,
-  type FaviconSort,
-  type TopicCandidateDirection,
-  type TopicCandidateSort,
-  type TopSiteDirection,
-  type TopSiteSort,
   type DeclaredOriginOs,
   type ExtractCollection,
-  type HistoryDirection,
-  type HistorySort,
-  type HistoryView,
   type IngestProgress,
   type IngestResult,
-  type MetadataDirection,
-  type MetadataSort,
-  type MetadataType,
   type SourceKind,
 } from "@forensix/core";
 import { startDashboardServer } from "@forensix/server";
 import { readFileSync } from "node:fs";
 import { dirname, join } from "node:path";
 import { fileURLToPath } from "node:url";
+
+import {
+  assertAllowedOptions,
+  enumOption,
+  hasOption,
+  integerOption,
+  optionValue,
+  optionValues,
+  parseArguments,
+  requiredCasePath,
+} from "./args.js";
+import { QUERY_COMMANDS } from "./query-commands.js";
 
 /**
  * Release-injected build identity. The release workflow
@@ -187,121 +159,6 @@ analyse exit codes:
   1  error    usage error, missing Case, or Working Copy integrity refusal
 `;
 
-type OptionValue = string | true;
-
-interface ParsedArguments {
-  readonly command: string | undefined;
-  readonly positionals: readonly string[];
-  readonly options: ReadonlyMap<string, readonly OptionValue[]>;
-}
-
-const FLAG_OPTIONS = new Set([
-  "--json",
-  "--include-tier-2",
-  "--csv",
-  "--include-secrets",
-  "--decrypt",
-  "--no-topic-classification",
-]);
-const VALUE_OPTIONS = new Set([
-  "--case",
-  "--port",
-  "--source-kind",
-  "--timezone",
-  "--origin-os",
-  "--view",
-  "--profile",
-  "--search",
-  "--commit-state",
-  "--transition",
-  "--from",
-  "--to",
-  "--host",
-  "--same-site",
-  "--state",
-  "--danger",
-  "--category",
-  "--kind",
-  "--type",
-  "--source",
-  "--record-type",
-  "--backend",
-  "--sort",
-  "--direction",
-  "--limit",
-  "--after",
-  "--out",
-  "--collection",
-  "--examiner",
-  "--extract",
-  "--key-material",
-  "--recipient-key",
-  "--label",
-]);
-const REPEATABLE_OPTIONS = new Set(["--profile", "--collection"]);
-
-function parseArguments(arguments_: readonly string[]): ParsedArguments {
-  const [command, ...remaining] = arguments_;
-  const positionals: string[] = [];
-  const options = new Map<string, OptionValue[]>();
-
-  for (let index = 0; index < remaining.length; index += 1) {
-    const argument = remaining[index];
-    if (argument === undefined) {
-      continue;
-    }
-    if (!argument.startsWith("--")) {
-      positionals.push(argument);
-      continue;
-    }
-    if (!FLAG_OPTIONS.has(argument) && !VALUE_OPTIONS.has(argument)) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        `Unknown option: ${argument}`,
-      );
-    }
-    const existing = options.get(argument);
-    if (existing !== undefined && !REPEATABLE_OPTIONS.has(argument)) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        `Option is repeated: ${argument}`,
-      );
-    }
-    if (FLAG_OPTIONS.has(argument)) {
-      options.set(argument, [true]);
-      continue;
-    }
-    const value = remaining[index + 1];
-    if (value === undefined || value.startsWith("--")) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        `Option ${argument} needs a value.`,
-      );
-    }
-    options.set(argument, [...(existing ?? []), value]);
-    index += 1;
-  }
-
-  return { command, positionals, options };
-}
-
-function hasOption(arguments_: ParsedArguments, option: string): boolean {
-  return arguments_.options.has(option);
-}
-
-function optionValues(arguments_: ParsedArguments, option: string): string[] {
-  return (arguments_.options.get(option) ?? []).filter(
-    (value): value is string => typeof value === "string",
-  );
-}
-
-function optionValue(
-  arguments_: ParsedArguments,
-  option: string,
-): string | undefined {
-  return optionValues(arguments_, option)[0];
-}
-
 function nodeVersionIsSupported(version: string): boolean {
   const actual = version.split(".").map((part) => Number(part));
   const minimum = MINIMUM_NODE_VERSION.split(".").map((part) => Number(part));
@@ -313,67 +170,6 @@ function nodeVersionIsSupported(version: string): boolean {
     }
   }
   return true;
-}
-
-function requiredCasePath(arguments_: ParsedArguments): string {
-  const casePath = optionValue(arguments_, "--case");
-  if (casePath === undefined) {
-    throw new ForensixError("INVALID_ARGUMENT", "Option --case is required.");
-  }
-  return casePath;
-}
-
-function assertAllowedOptions(
-  arguments_: ParsedArguments,
-  allowed: ReadonlySet<string>,
-): void {
-  const invalid = [...arguments_.options.keys()].filter(
-    (option) => !allowed.has(option),
-  );
-  if (invalid.length > 0) {
-    throw new ForensixError(
-      "INVALID_ARGUMENT",
-      `The ${arguments_.command ?? "unknown"} command has invalid options.`,
-      { invalid_options: invalid },
-    );
-  }
-}
-
-function enumOption<const Values extends readonly string[]>(
-  arguments_: ParsedArguments,
-  option: string,
-  values: Values,
-): Values[number] | undefined {
-  const value = optionValue(arguments_, option);
-  if (value === undefined) {
-    return undefined;
-  }
-  if (!values.includes(value)) {
-    throw new ForensixError(
-      "INVALID_ARGUMENT",
-      `Option ${option} has an unsupported value: ${value}`,
-      { option, value, allowed: values },
-    );
-  }
-  return value as Values[number];
-}
-
-function integerOption(
-  arguments_: ParsedArguments,
-  option: string,
-): number | undefined {
-  const value = optionValue(arguments_, option);
-  if (value === undefined) {
-    return undefined;
-  }
-  if (!/^[0-9]+$/.test(value)) {
-    throw new ForensixError(
-      "INVALID_ARGUMENT",
-      `Option ${option} must be an integer.`,
-      { option, value },
-    );
-  }
-  return Number(value);
 }
 
 async function consumeIngest(
@@ -503,597 +299,19 @@ async function run(arguments_: readonly string[]): Promise<number> {
     return ANALYSE_EXIT_CODES[result.exitState];
   }
 
-  if (parsed.command === "history") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--view",
-        "--profile",
-        "--search",
-        "--commit-state",
-        "--transition",
-        "--from",
-        "--to",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
+  const queryCommand =
+    parsed.command === undefined
+      ? undefined
+      : QUERY_COMMANDS.get(parsed.command);
+  if (queryCommand !== undefined) {
+    assertAllowedOptions(parsed, queryCommand.allowedOptions);
     if (parsed.positionals.length !== 0) {
       throw new ForensixError(
         "INVALID_ARGUMENT",
-        "The history command accepts no positional values.",
+        `The ${parsed.command} command accepts no positional values.`,
       );
     }
-    const result = queryHistory({
-      caseDirectory: requiredCasePath(parsed),
-      view: enumOption(parsed, "--view", [
-        "visits",
-        "activity",
-        "most-visited",
-        "durations",
-      ] as const) as HistoryView | undefined,
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      commitState: enumOption(parsed, "--commit-state", [
-        "committed",
-        "wal_resident",
-        "journal_resident",
-      ] as const) as CommitState | undefined,
-      transition: optionValue(parsed, "--transition"),
-      from: optionValue(parsed, "--from"),
-      to: optionValue(parsed, "--to"),
-      sort: enumOption(parsed, "--sort", [
-        "visit-time",
-        "local-time",
-        "url",
-        "duration",
-        "visit-count",
-        "profile",
-      ] as const) as HistorySort | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | HistoryDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-    return 0;
-  }
-
-  if (parsed.command === "cookies") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--profile",
-        "--search",
-        "--commit-state",
-        "--host",
-        "--same-site",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
-    if (parsed.positionals.length !== 0) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        "The cookies command accepts no positional values.",
-      );
-    }
-    const result = queryCookies({
-      caseDirectory: requiredCasePath(parsed),
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      commitState: enumOption(parsed, "--commit-state", [
-        "committed",
-        "wal_resident",
-        "journal_resident",
-      ] as const) as CommitState | undefined,
-      host: optionValue(parsed, "--host"),
-      sameSite: optionValue(parsed, "--same-site"),
-      sort: enumOption(parsed, "--sort", [
-        "host",
-        "name",
-        "creation-time",
-        "expires-time",
-        "last-access-time",
-        "profile",
-      ] as const) as CookieSort | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | CookieDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-    return 0;
-  }
-
-  if (parsed.command === "metadata") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--profile",
-        "--search",
-        "--commit-state",
-        "--type",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
-    if (parsed.positionals.length !== 0) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        "The metadata command accepts no positional values.",
-      );
-    }
-    const result = queryMetadata({
-      caseDirectory: requiredCasePath(parsed),
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      commitState: enumOption(parsed, "--commit-state", [
-        "committed",
-        "wal_resident",
-        "journal_resident",
-      ] as const) as CommitState | undefined,
-      type: enumOption(parsed, "--type", [
-        "browser_metadata",
-        "profile_metadata",
-      ] as const) as MetadataType | undefined,
-      sort: enumOption(parsed, "--sort", ["type", "profile"] as const) as
-        | MetadataSort
-        | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | MetadataDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-    return 0;
-  }
-
-  if (parsed.command === "bookmarks") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--profile",
-        "--search",
-        "--commit-state",
-        "--source",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
-    if (parsed.positionals.length !== 0) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        "The bookmarks command accepts no positional values.",
-      );
-    }
-    const result = queryBookmarks({
-      caseDirectory: requiredCasePath(parsed),
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      commitState: enumOption(parsed, "--commit-state", [
-        "committed",
-        "wal_resident",
-        "journal_resident",
-      ] as const) as CommitState | undefined,
-      source: enumOption(parsed, "--source", ["primary", "backup"] as const) as
-        | BookmarkSource
-        | undefined,
-      sort: enumOption(parsed, "--sort", [
-        "name",
-        "url",
-        "date-added",
-        "folder",
-        "profile",
-      ] as const) as BookmarkSort | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | BookmarkDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-    return 0;
-  }
-
-  if (parsed.command === "credentials") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--profile",
-        "--search",
-        "--commit-state",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
-    if (parsed.positionals.length !== 0) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        "The credentials command accepts no positional values.",
-      );
-    }
-    const result = queryCredentials({
-      caseDirectory: requiredCasePath(parsed),
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      commitState: enumOption(parsed, "--commit-state", [
-        "committed",
-        "wal_resident",
-        "journal_resident",
-      ] as const) as CommitState | undefined,
-      sort: enumOption(parsed, "--sort", [
-        "created-time",
-        "last-used-time",
-        "origin",
-        "username",
-        "profile",
-      ] as const) as CredentialSort | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | CredentialDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-    return 0;
-  }
-
-  if (parsed.command === "top-sites") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--profile",
-        "--search",
-        "--commit-state",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
-    if (parsed.positionals.length !== 0) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        "The top-sites command accepts no positional values.",
-      );
-    }
-    const result = queryTopSites({
-      caseDirectory: requiredCasePath(parsed),
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      commitState: enumOption(parsed, "--commit-state", [
-        "committed",
-        "wal_resident",
-        "journal_resident",
-      ] as const) as CommitState | undefined,
-      sort: enumOption(parsed, "--sort", [
-        "rank",
-        "url",
-        "title",
-        "profile",
-      ] as const) as TopSiteSort | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | TopSiteDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-    return 0;
-  }
-
-  if (parsed.command === "topic-candidates") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--profile",
-        "--search",
-        "--commit-state",
-        "--label",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
-    if (parsed.positionals.length !== 0) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        "The topic-candidates command accepts no positional values.",
-      );
-    }
-    const result = queryTopicCandidates({
-      caseDirectory: requiredCasePath(parsed),
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      commitState: enumOption(parsed, "--commit-state", [
-        "committed",
-        "wal_resident",
-        "journal_resident",
-      ] as const) as CommitState | undefined,
-      label: optionValue(parsed, "--label"),
-      sort: enumOption(parsed, "--sort", [
-        "rank",
-        "supporting",
-        "label",
-        "profile",
-      ] as const) as TopicCandidateSort | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | TopicCandidateDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-    return 0;
-  }
-
-  if (parsed.command === "favicons") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--profile",
-        "--search",
-        "--commit-state",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
-    if (parsed.positionals.length !== 0) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        "The favicons command accepts no positional values.",
-      );
-    }
-    const result = queryFavicons({
-      caseDirectory: requiredCasePath(parsed),
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      commitState: enumOption(parsed, "--commit-state", [
-        "committed",
-        "wal_resident",
-        "journal_resident",
-      ] as const) as CommitState | undefined,
-      sort: enumOption(parsed, "--sort", [
-        "icon-url",
-        "page-url",
-        "last-updated",
-        "width",
-        "profile",
-      ] as const) as FaviconSort | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | FaviconDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-    return 0;
-  }
-
-  if (parsed.command === "cache") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--profile",
-        "--search",
-        "--record-type",
-        "--backend",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
-    if (parsed.positionals.length !== 0) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        "The cache command accepts no positional values.",
-      );
-    }
-    const result = queryCache({
-      caseDirectory: requiredCasePath(parsed),
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      recordType: enumOption(parsed, "--record-type", [
-        "finding",
-        "candidate",
-      ] as const) as CacheRecordType | undefined,
-      backend: optionValue(parsed, "--backend"),
-      sort: enumOption(parsed, "--sort", [
-        "key",
-        "last-used",
-        "size",
-        "entry-hash",
-        "profile",
-      ] as const) as CacheSort | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | CacheDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-    return 0;
-  }
-
-  if (parsed.command === "autofill") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--profile",
-        "--search",
-        "--commit-state",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
-    if (parsed.positionals.length !== 0) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        "The autofill command accepts no positional values.",
-      );
-    }
-    const result = queryAutofill({
-      caseDirectory: requiredCasePath(parsed),
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      commitState: enumOption(parsed, "--commit-state", [
-        "committed",
-        "wal_resident",
-        "journal_resident",
-      ] as const) as CommitState | undefined,
-      sort: enumOption(parsed, "--sort", [
-        "created-time",
-        "last-used-time",
-        "field-name",
-        "value",
-        "profile",
-      ] as const) as AutofillSort | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | AutofillDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-    return 0;
-  }
-
-  if (parsed.command === "candidates") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--profile",
-        "--search",
-        "--category",
-        "--kind",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
-    if (parsed.positionals.length !== 0) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        "The candidates command accepts no positional values.",
-      );
-    }
-    const result = queryCandidates({
-      caseDirectory: requiredCasePath(parsed),
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      category: enumOption(parsed, "--category", [
-        "identity",
-        "behavior",
-      ] as const),
-      kind: optionValue(parsed, "--kind"),
-      sort: enumOption(parsed, "--sort", [
-        "rank",
-        "kind",
-        "supporting-count",
-        "value",
-        "profile",
-      ] as const) as CandidateSort | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | CandidateDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-    return 0;
-  }
-
-  if (parsed.command === "downloads") {
-    assertAllowedOptions(
-      parsed,
-      new Set([
-        "--case",
-        "--json",
-        "--profile",
-        "--search",
-        "--commit-state",
-        "--state",
-        "--danger",
-        "--sort",
-        "--direction",
-        "--limit",
-        "--after",
-      ]),
-    );
-    if (parsed.positionals.length !== 0) {
-      throw new ForensixError(
-        "INVALID_ARGUMENT",
-        "The downloads command accepts no positional values.",
-      );
-    }
-    const result = queryDownloads({
-      caseDirectory: requiredCasePath(parsed),
-      profiles: optionValues(parsed, "--profile"),
-      search: optionValue(parsed, "--search"),
-      commitState: enumOption(parsed, "--commit-state", [
-        "committed",
-        "wal_resident",
-        "journal_resident",
-      ] as const) as CommitState | undefined,
-      state: optionValue(parsed, "--state"),
-      dangerType: optionValue(parsed, "--danger"),
-      sort: enumOption(parsed, "--sort", [
-        "start-time",
-        "end-time",
-        "target-path",
-        "state",
-        "total-bytes",
-        "profile",
-      ] as const) as DownloadSort | undefined,
-      direction: enumOption(parsed, "--direction", ["asc", "desc"] as const) as
-        | DownloadDirection
-        | undefined,
-      limit: integerOption(parsed, "--limit"),
-      after: optionValue(parsed, "--after"),
-    });
+    const result = queryCommand.execute(parsed);
     process.stdout.write(`${JSON.stringify(result)}\n`);
     return 0;
   }
